Remove duplicate test (#37975)
