+++ src/rest/tests/rendering.js
@@ -1,10 +1,12 @@
 import { jest, test } from '@jest/globals'
 import { slug } from 'github-slugger'
+import { readdirSync, readFileSync } from 'fs'
+import path from 'path'
 
-import { getDOM } from '../../../tests/helpers/e2etest.js'
-import getRest from '#src/rest/lib/index.js'
+import { get, getDOM } from '../../../tests/helpers/e2etest.js'
 import { isApiVersioned, allVersions } from '../../../lib/all-versions.js'
 import { getDiffOpenAPIContentRest } from '../scripts/test-open-api-schema.js'
+import getRest, { REST_DATA_DIR, REST_SCHEMA_FILENAME } from '#src/rest/lib/index.js'
 
 describe('REST references docs', () => {
   jest.setTimeout(3 * 60 * 1000)
@@ -23,6 +25,40 @@ describe('REST references docs', () => {
       const schemaSlugs = checksRestOperations.map((operation) => slug(operation.title))
       expect(schemaSlugs.every((slug) => domH2Ids.includes(slug))).toBe(true)
     }
+  })
+
+  test('all category and subcategory REST pages render for free-pro-team', async () => {
+    // This currently just grabs the 'free-pro-team' schema, but ideally, we'd
+    // get a list of all categories across all versions.
+    const freeProTeamVersion = readdirSync(REST_DATA_DIR)
+      .filter((file) => file.startsWith('fpt'))
+      .shift()
+    const freeProTeamSchema = JSON.parse(
+      readFileSync(path.join(REST_DATA_DIR, freeProTeamVersion, REST_SCHEMA_FILENAME), 'utf8')
+    )
+
+    const restCategories = Object.entries(freeProTeamSchema)
+      .map(([key, subCategory]) => {
+        const subCategoryKeys = Object.keys(subCategory)
+        if (subCategoryKeys.length === 1) {
+          return key
+        } else {
+          return subCategoryKeys.map((elem) => `${key}/${elem}`)
+        }
+      })
+      .flat()
+
+    const statusCodes = await Promise.all(
+      restCategories.map(async (page) => {
+        const url = `/en/rest/${page}`
+        const res = await get(url)
+        return [url, res.statusCode]
+      })
+    )
+    for (const [url, status] of statusCodes) {
+      expect(status, url).toBe(200)
+    }
+    expect.assertions(restCategories.length)
   })
 
   test('test the latest version of the OpenAPI schema categories/subcategories to see if it matches the content/rest directory', async () => {
@@ -64,6 +100,36 @@ describe('REST references docs', () => {
         expect($('[data-testid=api-version-picker] button span').text()).toBe('')
       }
     }
+  })
+
+  describe('headings', () => {
+    test('rest pages do not render any headings with duplicate text', async () => {
+      const $ = await getDOM('/en/rest/actions/artifacts')
+      const headingText = $('body')
+        .find('h2, h3, h4, h5, h6')
+        .map((i, el) => $(el).text())
+        .get()
+        .sort()
+
+      const dupes = headingText.filter((item, index) => headingText.indexOf(item) !== index)
+
+      const message = `The following duplicate heading texts were found: ${dupes.join(', ')}`
+      expect(dupes.length, message).toBe(0)
+    })
+
+    test('rest pages do not render any headings with duplicate ids', async () => {
+      const $ = await getDOM('/en/rest/actions/artifacts')
+      const headingIDs = $('body')
+        .find('h2, h3, h4, h5, h6')
+        .map((i, el) => $(el).attr('id'))
+        .get()
+        .sort()
+
+      const dupes = headingIDs.filter((item, index) => headingIDs.indexOf(item) !== index)
+
+      const message = `The following duplicate heading IDs were found: ${dupes.join(', ')}`
+      expect(dupes.length, message).toBe(0)
+    })
   })
 })
